requirements/base.txt
Django~=4.2
psycopg[binary]==3.1.15

djangorestframework==3.14.0
django-cors-headers
django-filter

pillow==10.1.0
